+++ Firestore/Example/Tests/Integration/API/FIRDatabaseTests.mm
@@ -1316,6 +1316,47 @@ using firebase::firestore::util::TimerId;
                           NSException, @"The client has already been terminated.");
 }
 
+- (void)testTransactionGetAfterTerminationFail {
+  FIRDocumentReference *doc = [self documentRef];
+  FIRFirestore *firestore = doc.firestore;
+  XCTestExpectation *expectation1 = [self expectationWithDescription:@"TransactionBlockStart"];
+  XCTestExpectation *expectation2 = [self expectationWithDescription:@"FirestoreTerminated"];
+  XCTestExpectation *expectation3 = [self expectationWithDescription:@"TransactionBlockDone"];
+  __block NSError *capturedError = nil;
+
+  [firestore
+      runTransactionWithBlock:^id _Nullable(FIRTransaction *transaction, NSError **error) {
+        // Tell the test thread that it can now call `terminateWithCompletion`.
+        [expectation1 fulfill];
+        // Wait for `terminateWithCompletion` to complete.
+        [self awaitExpectation:expectation2];
+        // Call `[transaction getDocument]` to make sure that it correctly reports the error.
+        [transaction getDocument:doc error:error];
+        // Save the `NSError` into a variable on which the test thread will perform assertions.
+        capturedError = *error;
+        // Tell the test thread that it can now perform assertions on the captured `NSError` object.
+        [expectation3 fulfill];
+        return @"I should have failed since the transaction was terminated";
+      }
+                   completion:^(id, NSError *error){
+                   }];
+
+  // Wait for the transaction callback to start.
+  [self awaitExpectation:expectation1];
+
+  // Terminate the Firestore instance.
+  [firestore terminateWithCompletion:[self completionForExpectation:expectation2]];
+
+  // Wait for the transaction callback to "publish" the `NSError` from `[transaction getDocument]`.
+  [self awaitExpectation:expectation3];
+
+  // Verify that `[transaction getDocument]` set its `NSError` argument to a non-nil error.
+  XCTAssertNotNil(capturedError);
+  XCTAssertEqual(capturedError.code, FIRFirestoreErrorCodeFailedPrecondition);
+  XCTAssertEqualObjects(capturedError.userInfo[NSLocalizedDescriptionKey],
+                        @"The client has already been terminated.");
+}
+
 - (void)testMaintainsPersistenceAfterRestarting {
   FIRDocumentReference *doc = [self documentRef];
   FIRFirestore *firestore = doc.firestore;